Breadcrumb feature added.

+++ src/ReadXYZ/Twig/CacheTemplate.php
@@ -13,6 +13,8 @@ class CacheTemplate
     private Page   $page;
     private string $before = '';
     private string $after  = '';
+    private int    $beforeCount = -1;
+    private int    $afterCount = -1;
 
     public function __construct(string $errorMessage = '')
     {
@@ -31,17 +33,26 @@ class CacheTemplate
         return $files;
     }
 
-    private function clearTwigCache(): void
+    public function clearTwigCache(bool $force = false): void
     {
-        $cachePath = Util::getReadXyzSourcePath('cache');
-        $files = $this->recursiveGlob("$cachePath/*/*.php");
-        $count = count($files);
-        $this->before = "Attempting to delete $count Twig cache files.\n";
-        foreach ($files as $file) {unlink($file);}
-        $afterFiles = $this->recursiveGlob("$cachePath/*/*.php");
-        $afterCount = count($afterFiles);
-        $deletedCount = $count - $afterCount;
-        $this->after = "$deletedCount Twig cache files where successfully deleted.";
+        $cacheFlag = Util::getPublicPath('clear_cache.txt');
+        $flagExists = file_exists($cacheFlag) && !Util::isLocal();
+        if ($flagExists || $force) {
+            $cachePath = Util::getReadXyzSourcePath('cache');
+            $files = $this->recursiveGlob("$cachePath/*/*.php");
+            $count = count($files);
+            $this->before = "Attempting to delete $count Twig cache files.\n";
+            foreach ($files as $file) {unlink($file);}
+            $afterFiles = $this->recursiveGlob("$cachePath/*/*.php");
+            $afterCount = count($afterFiles);
+            $deletedCount = $count - $afterCount;
+            $this->after = "$deletedCount Twig cache files where successfully deleted.";
+            $this->afterCount = $afterCount;
+            $this->beforeCount = $count;
+        }
+        // only delete if local
+        if ($flagExists) unlink($cacheFlag);
+
     }
 // ======================== PUBLIC METHODS =====================
 
@@ -55,9 +66,12 @@ class CacheTemplate
             $this->page->addError($errorMessage);
         }
 
-
         $args = ['page' => $this->page, 'before' => $this->before, 'after' => $this->after];
         echo TwigFactory::getInstance()->renderTemplate('cache_status.html.twig', $args);
     }
 
+    public function returnCounts(): array
+    {
+        return ['before' => $this->beforeCount, 'after' => $this->afterCount, 'message' => $this->before . $this->after];
+    }
 }
